+++ arcade/application.py
@@ -30,6 +30,17 @@ MOUSE_BUTTON_RIGHT = 4
 
 _window: 'Window'
 
+__all__ = [
+    "get_screens",
+    "NoOpenGLException",
+    "Window",
+    "open_window",
+    "View",
+    "MOUSE_BUTTON_LEFT",
+    "MOUSE_BUTTON_MIDDLE",
+    "MOUSE_BUTTON_RIGHT"
+]
+
 
 def get_screens():
     """
@@ -137,7 +148,7 @@ class Window(pyglet.window.Window):
                 LOG.warning("Skipping antialiasing due missing hardware/driver support")
                 config = None
                 antialiasing = False
-        # If we still don't have a config 
+        # If we still don't have a config
         if not config:
             config = pyglet.gl.Config(
                 major_version=gl_version[0],