+++ arcade/experimental/atlas_replace_image.py
@@ -81,7 +81,7 @@ class AtlasReplaceImage(arcade.Window):
 
     def on_update(self, delta_time: float):
         self.elapsed_time += delta_time
-        # Change textures 
+        # Change textures
         if self.elapsed_time > 1.0:
             # Replace the internal images. They all have the same size (required)
             self.sprite_1.texture.image = next(self.texture_cycle).image