+++ arcade/physics_engines.py
@@ -14,6 +14,11 @@ from arcade import (
 )
 from arcade.math import get_distance
 
+__all__ = [
+    "PhysicsEngineSimple",
+    "PhysicsEnginePlatformer"
+]
+
 
 def _circular_check(player: Sprite, walls: List[SpriteList]):
     """