+++ tests/unit/color/test_color_type.py
@@ -2,6 +2,7 @@ import math
 from copy import deepcopy
 from itertools import product
 from typing import Iterable, Callable, Tuple
+from unittest.mock import Mock
 
 import pytest
 
@@ -13,6 +14,10 @@ from arcade.types import Color
 OK_NORMALIZED = (0.0, 1.0)
 BAD_NORMALIZED = (-0.01, 1.01)
 MIXED_NORMALIZED = OK_NORMALIZED + BAD_NORMALIZED
+
+
+class ColorSubclass(Color):
+    pass
 
 
 def at_least_one_in(i: Iterable) -> Callable[[Iterable], bool]:
@@ -44,6 +49,11 @@ def test_color_from_iterable_returns_same_color():
         assert Color.from_iterable(color) is color
 
 
+def test_color_from_iterable_inheritance():
+    color = ColorSubclass.from_iterable(colors.AO)
+    assert isinstance(color, ColorSubclass)
+
+
 def test_color_from_uint24():
     assert Color.from_uint24(0xFFFFFF) == (255, 255, 255, 255)
     assert Color.from_uint24((1 << 16) + (2 << 8) + 3) == (1, 2, 3, 255)
@@ -55,6 +65,11 @@ def test_color_from_uint24():
         Color.from_uint24("moo")
 
 
+def test_color_from_uint24_inheritance():
+    color = ColorSubclass.from_uint24(0xFFFFFF)
+    assert isinstance(color, ColorSubclass)
+
+
 def test_color_from_uint32():
     assert Color.from_uint32(4294967295) == (255, 255, 255, 255)
     assert Color.from_uint32((1 << 24) + (2 << 16) + (3 << 8) + 4) == (1, 2, 3, 4)
@@ -63,6 +78,11 @@ def test_color_from_uint32():
 
     with pytest.raises(TypeError):
         Color.from_uint32("bad")
+
+
+def test_color_from_uint32_inheritance():
+    color = ColorSubclass.from_uint32(0xFFFFFFFF)
+    assert isinstance(color, ColorSubclass)
 
 
 def test_color_from_normalized():
@@ -95,6 +115,11 @@ def test_color_from_normalized():
             Color.from_normalized(bad_rgba_channels)
 
 
+def test_from_normalized_inheritance():
+    color = ColorSubclass.from_normalized((1.0, 1.0, 1.0, 1.0))
+    assert isinstance(color, ColorSubclass)
+
+
 def test_color_from_gray():
     OK_255 = (0, 255)
     BAD_255 = (-1, 256)
@@ -116,6 +141,11 @@ def test_color_from_gray():
     for bad_arg in BAD_255:
         with pytest.raises(ValueError):
             Color.from_gray(bad_arg)
+
+
+def test_color_from_gray_inheritance():
+    color = ColorSubclass.from_gray(255, a=255)
+    assert isinstance(color, ColorSubclass)
 
 
 def test_color_from_hex_string():
@@ -148,6 +178,11 @@ def test_color_from_hex_string():
             Color.from_hex_string(bad_value)
 
 
+def test_color_from_hex_string_inheritance():
+    color = ColorSubclass.from_hex_string("#fff")
+    assert isinstance(color, ColorSubclass)
+
+
 def test_color_normalized_property():
     assert colors.BLACK.normalized == (0.0, 0.0, 0.0, 1.0)
     assert colors.WHITE.normalized == (1.0, 1.0, 1.0, 1.0)
@@ -155,6 +190,38 @@ def test_color_normalized_property():
     assert colors.GRAY.normalized == (128 / 255, 128 / 255, 128 / 255, 1.0)
 
 
-def test_deepcopy_color():
+def test_deepcopy_color_values():
     expected_color = Color(255, 255, 255, 255)
     assert deepcopy(expected_color) == expected_color
+
+
+def test_deepcopy_color_inheritance():
+    color_subclass_instance = ColorSubclass(255, 255, 255, a=255)
+    deep = deepcopy(color_subclass_instance)
+    assert isinstance(deep, ColorSubclass)
+
+
+RANDINT_RETURN_RESULT = 128
+
+
+@pytest.fixture
+def randint_is_constant(monkeypatch):
+    monkeypatch.setattr('random.randint', Mock(return_value=RANDINT_RETURN_RESULT))
+
+
+def test_color_random(randint_is_constant):
+
+    for combo in product((None, 0), repeat=4):
+        color = Color.random(*combo)
+        for channel_value, channel_arg in zip(color, combo):
+            if channel_arg is None:
+                expected = RANDINT_RETURN_RESULT
+            else:
+                expected = 0
+
+            assert channel_value == expected
+
+
+def test_color_random_inheritance(randint_is_constant):
+    color = ColorSubclass.random()
+    assert isinstance(color, ColorSubclass)