+++ doc/programming_guide/release_notes.rst
@@ -47,6 +47,7 @@ API in a way that is not compatible with how it was used in 2.6.
   are also using any custom :py:class:`~arcade.TextureAtlas`.
 * The GUI package has been changed significantly.
 * Buffered shapes (shape list items) have been moved to their own sub-module.
+* `use_spatial_hash` parameter for `SpriteList` and `TileMap` is now a `bool` instead of `Optional[bool]`
 
 Featured Updates
 ~~~~~~~~~~~~~~~~
@@ -105,10 +106,12 @@ Changes
         * ``UIWidget.padding_bottom``
         * ``UIWidget.padding_left``
     * Update and add example code.
+    * Iterable (providing direct children)
 
   * New widgets:
 
     * :py:class:`~arcade.gui.widgets.dropdown.UIDropdown`
+    * :py:class:`~arcade.gui.widgets.image.UIImage`
     * :py:class:`~arcade.gui.widgets.slider.UISlider`
     * :py:class:`~arcade.gui.widgets.constructs.UIButtonRow` (`PR1580 <https://github.com/pythonarcade/arcade/pull/1580>`_ and `PR1253 <https://github.com/pythonarcade/arcade/pull/1253>`_)
 
@@ -184,6 +187,8 @@ Changes
   * Uniforms are now set using ``glProgramUniform`` instead of ``glUniform``
     when the extension is available.
   * Fixed many implicit type conversions in the shader code for wider support.
+  * Added ``front_face`` property on the context for configuring front face winding order of triangles
+  * Added ``cull_face`` property on the context for configuring what triangle face to cull
 
 * :py:class:`~arcade.tilemap.TileMap`
 
@@ -2349,4 +2354,3 @@ Enhancements
 * `Issue 131 <https://github.com/pvcraven/arcade/issues/131>`_: Add example code on how to do full-screen games
 * `Issue 113 <https://github.com/pvcraven/arcade/issues/113>`_: Add example code showing enemy turning around when hitting a wall
 * `Issue 67 <https://github.com/pvcraven/arcade/issues/67>`_: Improved support and documentation for joystick/game controllers
-